A live component was created for Board.
