+++ lib/matchum_web/live/game_live.ex
@@ -2,6 +2,7 @@ defmodule MatchumWeb.GameLive do
   use MatchumWeb, :live_view
   alias Matchum.Game
   import Phoenix.HTML
+
   @numbers %{
     1 => "blue",
     2 => "green",
@@ -10,7 +11,8 @@ defmodule MatchumWeb.GameLive do
     5 => "salmon",
     6 => "orange",
     7 => "black",
-    8 => "red"}
+    8 => "red"
+  }
 
   @impl true
   def mount(_params, _session, socket) do
@@ -29,20 +31,17 @@ defmodule MatchumWeb.GameLive do
   end
 
   def render(assigns) do
-      # <%= inspect @board %>
+    # <%= inspect @board %>
     ~L"""
     <pre>
-    <%= raw(render_board(@game)) %>
+    <%= live_component @socket, MatchumWeb.Board, game: @game %>
     <%= raw(render_buttons()) %>
+
     Proposed Guess: <%= raw(render_proposed_guess(@proposed_guess)) %>
     <button phx-click="submit" <%= if !valid_guess?(@proposed_guess), do: "disabled" %> > Submit Guess </button>
     <button phx-click="clear"> Clear </button>
     </pre>
     """
-  end
-
-  defp render_board(game) do
-    render_rows(game.rows) <> render_status(game.status)
   end
 
   defp render_buttons() do
@@ -60,34 +59,12 @@ defmodule MatchumWeb.GameLive do
     end
   end
 
-  defp render_rows(rows) do
-    for row <- rows do
-      """
-      <p> #{render_pegs(row.guess)} </p>
-      <p> #{inspect(row.score.reds)} #{inspect(row.score.whites)} </p>
-      """
-    end
-    |> Enum.join()
-  end
-
-  defp render_pegs(guess) do
-    for peg <- guess do
-      ~s[<span style="color: #{@numbers[peg]}"> (⌐■_■) : #{peg} </span>]
-    end
-  end
-
-  defp render_status(status) do
-    """
-    <div> #{status} </div>
-    """
-  end
-
   def handle_event("add_item", %{"code" => code, "color" => _color}, socket) do
     {:noreply, add_item(socket, String.to_integer(code))}
   end
 
   def handle_event("submit", _value, socket) do
-    {:noreply, socket |> append_guess |> clear |> build_game }
+    {:noreply, socket |> append_guess |> clear |> build_game}
   end
 
   def handle_event("clear", _meta, socket) do
@@ -103,11 +80,11 @@ defmodule MatchumWeb.GameLive do
     assign(socket, board: updated_board)
   end
 
-  defp add_item(%{assigns: %{proposed_guess: guess}}=socket, _color) when length(guess) == 4, do: socket
+  defp add_item(%{assigns: %{proposed_guess: guess}} = socket, _color) when length(guess) == 4,
+    do: socket
+
   defp add_item(socket, color) do
     updated_proposed_guess = socket.assigns.proposed_guess ++ [color]
     assign(socket, proposed_guess: updated_proposed_guess)
   end
-
-
 end
